Update bash_prompt colors for git status
As a temporary fix until I can figure out to abstract color theme
values in a separate value to be read in by bash_prompt script, hardcode
the values in the prompt_git() function for the various statuses.  One
thing to note, the order status is important in order to get the correct
color for a particular status.

Issue: #9